Remove stripe onboarding
