+++ client/my-sites/earn/memberships/products-list.tsx
@@ -3,7 +3,7 @@ import {
 	FEATURE_PREMIUM_CONTENT_CONTAINER,
 	FEATURE_RECURRING_PAYMENTS,
 } from '@automattic/calypso-products';
-import { Badge, Button, CompactCard, Gridicon } from '@automattic/components';
+import { Badge, Button, Card, CompactCard, Gridicon } from '@automattic/components';
 import formatCurrency from '@automattic/format-currency';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useState } from 'react';
@@ -17,7 +17,6 @@ import SectionHeader from 'calypso/components/section-header';
 import { useDispatch, useSelector } from 'calypso/state';
 import { bumpStat } from 'calypso/state/analytics/actions';
 import { getProductsForSiteId } from 'calypso/state/memberships/product-list/selectors';
-import { getConnectedAccountIdForSiteId } from 'calypso/state/memberships/settings/selectors';
 import getFeaturesBySiteId from 'calypso/state/selectors/get-site-features';
 import siteHasFeature from 'calypso/state/selectors/site-has-feature';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
@@ -46,9 +45,8 @@ function ProductsList() {
 	const features = useSelector( ( state ) => getFeaturesBySiteId( state, site?.ID ) );
 	const hasLoadedFeatures = features?.active.length > 0;
 	const products: Product[] = useSelector( ( state ) => getProductsForSiteId( state, site?.ID ) );
-	const connectedAccountId = useSelector( ( state ) =>
-		getConnectedAccountIdForSiteId( state, site?.ID )
-	);
+	const hasProducts = products.length > 0;
+
 	const hasDonationsFeature = useSelector( ( state ) =>
 		siteHasFeature( state, site?.ID ?? null, FEATURE_DONATIONS )
 	);
@@ -160,12 +158,21 @@ function ProductsList() {
 				/>
 			) }
 
-			{ hasLoadedFeatures && hasStripeFeature && connectedAccountId && (
+			{ hasLoadedFeatures && hasStripeFeature && (
 				<SectionHeader label={ translate( 'Manage plans' ) }>
 					<Button primary compact onClick={ () => openAddEditDialog() }>
 						{ translate( 'Add a new payment plan' ) }
 					</Button>
 				</SectionHeader>
+			) }
+			{ hasLoadedFeatures && hasStripeFeature && ! hasProducts && (
+				<Card className="memberships__products-card">
+					<div className="memberships__products-card-content">
+						<div className="memberships__products-card-title">
+							{ translate( 'Set up a one-time offer or recurring payments plan.' ) }
+						</div>
+					</div>
+				</Card>
 			) }
 			{ hasLoadedFeatures &&
 				products
@@ -216,7 +223,7 @@ function ProductsList() {
 							</CompactCard>
 						);
 					} ) }
-			{ hasLoadedFeatures && showAddEditDialog && hasStripeFeature && connectedAccountId && (
+			{ hasLoadedFeatures && showAddEditDialog && hasStripeFeature && (
 				<RecurringPaymentsPlanAddEditModal
 					closeDialog={ closeDialog }
 					product={ Object.assign( product ?? {}, {
