+++ client/my-sites/earn/memberships/section.tsx
@@ -1,27 +1,16 @@
-import { Card, Button, Dialog, Gridicon } from '@automattic/components';
-import { localizeUrl } from '@automattic/i18n-utils';
 import { useTranslate } from 'i18n-calypso';
-import { useState, useEffect, useCallback, ReactNode } from 'react';
-import paymentsImage from 'calypso/assets/images/earn/payments-illustration.svg';
+import { useEffect, useCallback } from 'react';
 import QueryMembershipsSettings from 'calypso/components/data/query-memberships-settings';
 import { LoadingEllipsis } from 'calypso/components/loading-ellipsis';
 import Notice from 'calypso/components/notice';
 import NoticeAction from 'calypso/components/notice/notice-action';
-import SectionHeader from 'calypso/components/section-header';
-import { preventWidows } from 'calypso/lib/formatting';
-import { userCan } from 'calypso/lib/site/utils';
-import { useDispatch, useSelector } from 'calypso/state';
-import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import { getEarningsWithDefaultsForSiteId } from 'calypso/state/memberships/earnings/selectors';
+import { useSelector } from 'calypso/state';
 import { getProductsForSiteId } from 'calypso/state/memberships/product-list/selectors';
-import { requestDisconnectSiteStripeAccount } from 'calypso/state/memberships/settings/actions';
 import {
 	getConnectedAccountIdForSiteId,
-	getConnectedAccountDescriptionForSiteId,
 	getConnectUrlForSiteId,
 } from 'calypso/state/memberships/settings/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
-import CommissionFees from '../components/commission-fees';
 import { Query } from '../types';
 import {
 	ADD_TIER_PLAN_HASH,
@@ -30,6 +19,7 @@ import {
 } from './constants';
 import ProductList from './products-list';
 import './style.scss';
+import Settings from './settings';
 
 type MembershipsSectionProps = {
 	query?: Query;
@@ -37,9 +27,8 @@ type MembershipsSectionProps = {
 
 function MembershipsSection( { query }: MembershipsSectionProps ) {
 	const translate = useTranslate();
-	const dispatch = useDispatch();
+
 	const source = getSource();
-	const [ disconnectedConnectedAccountId, setDisconnectedConnectedAccountId ] = useState( null );
 
 	const site = useSelector( ( state ) => getSelectedSite( state ) );
 
@@ -49,14 +38,7 @@ function MembershipsSection( { query }: MembershipsSectionProps ) {
 
 	const products = useSelector( ( state ) => getProductsForSiteId( state, site?.ID ) );
 
-	const connectedAccountDescription = useSelector( ( state ) =>
-		getConnectedAccountDescriptionForSiteId( state, site?.ID )
-	);
 	const connectUrl: string = useSelector( ( state ) => getConnectUrlForSiteId( state, site?.ID ) );
-
-	const { commission } = useSelector( ( state ) =>
-		getEarningsWithDefaultsForSiteId( state, site?.ID )
-	);
 
 	const navigateToLaunchpad = useCallback( () => {
 		const shouldGoToLaunchpad = query?.stripe_connect_success === 'launchpad';
@@ -66,88 +48,9 @@ function MembershipsSection( { query }: MembershipsSectionProps ) {
 		}
 	}, [ query, site ] );
 
-	function onCloseDisconnectStripeAccount( reason: string | undefined ) {
-		if ( reason === 'disconnect' ) {
-			dispatch(
-				requestDisconnectSiteStripeAccount(
-					site?.ID,
-					connectedAccountId,
-					translate( 'Please wait, disconnecting Stripe\u2026' ),
-					translate( 'Stripe account is disconnected.' )
-				)
-			);
-		}
-		setDisconnectedConnectedAccountId( null );
-	}
-
-	function renderSettings() {
-		return (
-			<div>
-				<SectionHeader label={ translate( 'Settings' ) } />
-				<Card
-					onClick={ () => setDisconnectedConnectedAccountId( connectedAccountId ) }
-					className="memberships__settings-link"
-				>
-					<div className="memberships__module-plans-content">
-						<div className="memberships__module-plans-icon">
-							<Gridicon size={ 24 } icon="link-break" />
-						</div>
-						<div>
-							<div className="memberships__module-settings-title">
-								{ translate( 'Disconnect Stripe Account' ) }
-							</div>
-							{ connectedAccountDescription ? (
-								<div className="memberships__module-settings-description">
-									{ translate( 'Connected to %(connectedAccountDescription)s', {
-										args: {
-											connectedAccountDescription: connectedAccountDescription,
-										},
-									} ) }
-								</div>
-							) : null }
-						</div>
-					</div>
-				</Card>
-				<Dialog
-					isVisible={ !! disconnectedConnectedAccountId }
-					buttons={ [
-						{
-							label: translate( 'Cancel' ),
-							action: 'cancel',
-						},
-						{
-							label: translate( 'Disconnect Payments from Stripe' ),
-							isPrimary: true,
-							action: 'disconnect',
-						},
-					] }
-					onClose={ onCloseDisconnectStripeAccount }
-				>
-					<h1>{ translate( 'Confirmation' ) }</h1>
-					<p>{ translate( 'Do you want to disconnect Payments from your Stripe account?' ) }</p>
-					<Notice
-						text={ translate(
-							'Once you disconnect Payments from Stripe, new subscribers won’t be able to sign up and existing subscriptions will stop working.'
-						) }
-						showDismiss={ false }
-					/>
-				</Dialog>
-			</div>
-		);
-	}
-
-	function renderStripeConnected() {
-		return (
-			<div>
-				{ renderNotices() }
-				<ProductList />
-				{ renderSettings() }
-			</div>
-		);
-	}
-
 	function renderNotices() {
 		const stripe_connect_success = query?.stripe_connect_success;
+		const stripe_connect_canclled = Boolean( query?.stripe_connect_success );
 
 		if ( stripe_connect_success === 'earn' ) {
 			const siteHasPlans = products.length !== 0;
@@ -162,13 +65,11 @@ function MembershipsSection( { query }: MembershipsSectionProps ) {
 			}
 
 			return (
-				! siteHasPlans && (
-					<Notice status="is-success" showDismiss={ false } text={ congratsText }>
-						<NoticeAction href={ `/earn/payments/${ site?.slug }` } icon="create">
-							{ translate( 'Add a payment plan' ) }
-						</NoticeAction>
-					</Notice>
-				)
+				<Notice status="is-success" showDismiss={ false } text={ congratsText }>
+					<NoticeAction href={ `/earn/payments/${ site?.slug }` } icon="create">
+						{ translate( 'Add a payment plan' ) }
+					</NoticeAction>
+				</Notice>
 			);
 		}
 
@@ -192,142 +93,23 @@ function MembershipsSection( { query }: MembershipsSectionProps ) {
 			);
 		}
 
+		if ( stripe_connect_canclled ) {
+			return (
+				<Notice
+					showDismiss={ false }
+					text={ translate(
+						'The attempt to connect to Stripe has been cancelled. You can connect again at any time.'
+					) }
+				/>
+			);
+		}
+
 		return null;
-	}
-
-	function renderOnboarding( cta: ReactNode, intro?: ReactNode | null ) {
-		return (
-			<Card>
-				<div className="memberships__onboarding-wrapper">
-					<div className="memberships__onboarding-column-info">
-						<h2 className="memberships__onboarding-header">
-							{ preventWidows( translate( 'Accept payments on your website.' ) ) }
-						</h2>
-						<p className="memberships__onboarding-paragraph">
-							{ preventWidows(
-								translate(
-									'Our payments blocks make it easy to add a buy button for digital goods or services, collect donations via a form, or limit access for specific content to subscribers-only.'
-								)
-							) }
-						</p>
-						<p className="memberships__onboarding-paragraph">
-							{ preventWidows(
-								translate(
-									'The Payment Button, Donations Form, and Premium Content blocks all require you to first connect your bank account details with our secure payment processor, Stripe.',
-									{
-										components: {
-											link: (
-												<a
-													href={ localizeUrl(
-														'https://wordpress.com/support/wordpress-editor/blocks/payments/#setting-up-payments'
-													) }
-												/>
-											),
-										},
-									}
-								)
-							) }
-						</p>
-						{ intro ? <p className="memberships__onboarding-paragraph">{ intro }</p> : null }
-						<p className="memberships__onboarding-paragraph">{ cta }</p>
-						<p className="memberships__onboarding-paragraph memberships__onboarding-paragraph-disclaimer">
-							<em>
-								{ preventWidows(
-									translate(
-										'All credit and debit card payments made through these blocks are securely and seamlessly processed by Stripe.'
-									)
-								) }
-							</em>
-						</p>
-					</div>
-					<div className="memberships__onboarding-column-image">
-						<img src={ paymentsImage } aria-hidden="true" alt="" />
-					</div>
-				</div>
-				<div className="memberships__onboarding-benefits">
-					<div>
-						<h3>{ translate( 'No plugin required' ) }</h3>
-						{ preventWidows(
-							translate(
-								'No additional installs or purchases. Simply connect your banking details with our payment processor, Stripe, and insert a block to get started.'
-							)
-						) }
-					</div>
-					<div>
-						<h3>{ translate( 'One-time and recurring options' ) }</h3>
-						{ preventWidows(
-							translate(
-								'Accept one-time, monthly, and yearly payments from your visitors. This is perfect for a single purchase or tip — or a recurring donation, membership fee, or subscription.'
-							)
-						) }
-					</div>
-					<div>
-						<h3>{ translate( 'Simple fees structure' ) }</h3>
-						<p>
-							<CommissionFees commission={ commission } siteSlug={ site?.slug } />
-						</p>
-						<p>{ preventWidows( translate( 'No fixed monthly or annual fees charged.' ) ) }</p>
-					</div>
-					<div>
-						<h3>{ translate( 'Join thousands of others' ) }</h3>
-						{ preventWidows(
-							translate(
-								'Sites that actively promoted their businesses and causes on social media, email, and other platforms have collected tens of thousands of dollars through these blocks.'
-							)
-						) }
-					</div>
-				</div>
-			</Card>
-		);
-	}
-
-	function renderConnectStripe() {
-		return (
-			<div>
-				{ query?.stripe_connect_cancelled && (
-					<Notice
-						showDismiss={ false }
-						text={ translate(
-							'The attempt to connect to Stripe has been cancelled. You can connect again at any time.'
-						) }
-					/>
-				) }
-				{ renderOnboarding(
-					<Button
-						primary={ true }
-						href={ connectUrl }
-						onClick={ () =>
-							dispatch( recordTracksEvent( 'calypso_memberships_stripe_connect_click' ) )
-						}
-					>
-						{ translate( 'Connect Stripe to Get Started' ) }{ ' ' }
-						<Gridicon size={ 18 } icon="external" />
-					</Button>,
-					connectedAccountDescription
-						? translate( 'Previously connected to Stripe account %(connectedAccountDescription)s', {
-								args: {
-									connectedAccountDescription: connectedAccountDescription,
-								},
-						  } )
-						: null
-				) }
-			</div>
-		);
 	}
 
 	useEffect( () => {
 		navigateToLaunchpad();
 	}, [ navigateToLaunchpad ] );
-
-	if ( ! userCan( 'manage_options', site ) ) {
-		return renderOnboarding(
-			<Notice
-				status="is-warning"
-				text={ translate( 'Only site administrators can edit Payments settings.' ) }
-				showDismiss={ false }
-			/>
-		);
-	}
 
 	if ( ! site ) {
 		return <LoadingEllipsis />;
@@ -336,14 +118,16 @@ function MembershipsSection( { query }: MembershipsSectionProps ) {
 	return (
 		<div>
 			<QueryMembershipsSettings siteId={ site.ID } source={ source } />
-			{ connectedAccountId && renderStripeConnected() }
-			{ connectUrl && renderConnectStripe() }
-
 			{ ! connectedAccountId && ! connectUrl && (
 				<div className="earn__payments-loading">
 					<LoadingEllipsis />
 				</div>
 			) }
+			<div>
+				{ renderNotices() }
+				<ProductList />
+				<Settings />
+			</div>
 		</div>
 	);
 }
